Full cycle submit url, fetch, extract links and crawl the site using nats
You can go to `/add-site`
to submit a new sute to be cralwed
The url is sent to a nats server, then a fetcher worker pickes it up
and works with it, it passes another message to nats and then an
extractor worker picks that message up.

This runs as a loop until all pages have been cralwed.

Right now we don't crawl links to other domains.

There is also `/index-status` which shows you the current number
of pages fetched and parsed. In the future we will add more info here.

the webapp folder has the front end code, `main.go` as well as html and build files

You need to run `grunt serve` while in dev mode and the run main.go.